Realtime

+++ desktop/server/helpers/remote.js
@@ -4,7 +4,6 @@ const fs = require("fs");
 const path = require("path");
 const mkdirp = require("mkdirp");
 const request = require("request-promise-native");
-const emitter = require('../helpers/emitter').emitter;
 const { add: errorLogAdd } = require("../models/log-error");
 const { eventLogModel, types: eventType } = require("../models/log-event");
 const { nodeModel } = require("../models/node");
@@ -12,9 +11,6 @@ const token = require("./token");
 const _base = require("./syncers/_base");
 const Utimes = require('@ronomon/utimes');
 const _path = require('./path');
-
-// Logger
-const { logger } = require("./logger");
 
 /**
  * @param object params
@@ -36,6 +32,8 @@ exports.getNode = async params => {
 
   var options = {
     method: "GET",
+    pool: { maxSockets: 1 },
+    resolveWithFullResponse: true,
     url: `${account.instance_url}/alfresco/api/-default-/public/alfresco/versions/1/nodes/${record.node_id}?include=path`,
     headers: {
       Connection: "keep-alive",
@@ -43,24 +41,14 @@ exports.getNode = async params => {
     }
   };
 
-  request(options)
-    .then(response => {
-      emitter.emit('getNode' + record.node_id, {
-        account,
-        record,
-        statusCode: 200,
-        response: JSON.parse(String(response))
-      });
-    })
-    .catch(error => {
-      emitter.emit('getNode' + record.node_id, {
-        account,
-        record,
-        statusCode: error.statusCode,
-        response: {}
-      });
-      errorLogAdd(account.id, error, `${__filename}/getNode/${record.node_id}`);
-    });
+  try {
+    return await request(options);
+  } catch (error) {
+    console.log('error.message', error.message);
+    errorLogAdd(account.id, error, `${__filename}/getNode/${record.node_id}`);
+    error = JSON.parse(error.error);
+    return error.error;
+  }
 };
 
 /**
@@ -97,7 +85,7 @@ exports.getChildren = async params => {
   try {
     let response = await request(options)
       .on('error', function (e) {
-        console.error(e);
+        console.error('######ON ERROR#####', e);
         return;
       });
     return JSON.parse(response);
@@ -117,9 +105,14 @@ exports.deleteServerNode = async params => {
   let account = params.account;
   let record = params.record;
 
+  if (!account || !record) {
+    return;
+  }
+
   var options = {
-    resolveWithFullResponse: true,
     method: "DELETE",
+    pool: { maxSockets: 1 },
+    resolveWithFullResponse: true,
     url:
       account.instance_url +
       "/alfresco/api/-default-/public/alfresco/versions/1/nodes/" +
@@ -184,7 +177,7 @@ exports.deleteServerNode = async params => {
         }
       });
     } else {
-      errorLogAdd(account.id, error, `${__filename}/getServerNode`);
+      errorLogAdd(account.id, error, `${__filename}/deleteServerNode`);
     }
   }
 };
@@ -202,6 +195,7 @@ exports.download = async params => {
 
   var options = {
     method: "GET",
+    pool: { maxSockets: 1 },
     url:
       account.instance_url +
       "/alfresco/api/-default-/public/alfresco/versions/1/nodes/" +
@@ -275,27 +269,29 @@ exports.download = async params => {
           const atime = undefined;
 
           setTimeout(() => {
-            Utimes.utimes(`${destinationPath}`, btime, mtime, atime, async () => {
-
-              // set download progress to false
-              await nodeModel.update({
-                download_in_progress: 0,
-                file_update_at: mtime
-              }, {
-                  where: {
-                    account_id: account.id,
-                    file_path: _path.toUnix(destinationPath)
-                  }
-                });
-
-              // Add an event log
-              await eventLogModel.create({
-                account_id: account.id,
-                type: eventType.DOWNLOAD_FILE,
-                description: `Downloaded File: ${destinationPath} from ${account.instance_url}`,
-              });
-            });
+            Utimes.utimes(`${destinationPath}`, btime, mtime, atime, async () => { });
           }, 0);
+
+          // set download progress to false
+          await nodeModel.update({
+            download_in_progress: 0,
+            file_update_at: mtime
+          }, {
+              where: {
+                account_id: account.id,
+                file_path: _path.toUnix(destinationPath)
+              }
+            });
+
+          console.log(`Downloaded File: ${destinationPath} from ${account.instance_url}`);
+
+          // Add an event log
+          await eventLogModel.create({
+            account_id: account.id,
+            type: eventType.DOWNLOAD_FILE,
+            description: `Downloaded File: ${destinationPath} from ${account.instance_url}`,
+          });
+
         }
       })
       .pipe(fs.createWriteStream(destinationPath));
@@ -434,6 +430,7 @@ exports.upload = async params => {
     let relativePath = path.dirname(filePath).split('documentLibrary')[1].replace(/^\/|\/$/g, '');
 
     options = {
+      pool: { maxSockets: 1 },
       resolveWithFullResponse: true,
       method: "POST",
       url: `${
@@ -454,65 +451,64 @@ exports.upload = async params => {
       }
     };
 
-    request(options)
-      .then(async (response) => {
-        response = JSON.parse(response.body);
+    try {
+      let response = await request(options)
+      response = JSON.parse(response.body);
 
-        // Update the time meta properties of the downloaded file
-        const btime = _base.convertToUTC(response.entry.createdAt);
-        const mtime = _base.convertToUTC(response.entry.modifiedAt);
-        const atime = undefined;
+      // Update the time meta properties of the downloaded file
+      const btime = _base.convertToUTC(response.entry.createdAt);
+      const mtime = _base.convertToUTC(response.entry.modifiedAt);
+      const atime = undefined;
 
-        setTimeout(() => {
-          Utimes.utimes(`${filePath}`, btime, mtime, atime, async () => { });
-        }, 0); //end setTimeout
+      setTimeout(() => {
+        Utimes.utimes(`${filePath}`, btime, mtime, atime, async () => { });
+      }, 0); //end setTimeout
 
-        // Add a record in the db
-        await nodeModel.create({
-          account_id: account.id,
-          site_id: watcher.site_id,
-          node_id: response.entry.id,
-          remote_folder_path: response.entry.path.name,
-          file_name: path.basename(filePath),
-          file_path: _path.toUnix(filePath),
-          local_folder_path: path.dirname(filePath),
-          file_update_at: _base.convertToUTC(response.entry.modifiedAt),
-          last_uploaded_at: _base.getCurrentTime(),
-          last_downloaded_at: 0,
-          is_folder: false,
-          is_file: true,
-          download_in_progress: 0,
-          upload_in_progress: 0,
-        });
+      // Add a record in the db
+      await nodeModel.create({
+        account_id: account.id,
+        site_id: watcher.site_id,
+        node_id: response.entry.id,
+        remote_folder_path: response.entry.path.name,
+        file_name: path.basename(filePath),
+        file_path: _path.toUnix(filePath),
+        local_folder_path: path.dirname(filePath),
+        file_update_at: _base.convertToUTC(response.entry.modifiedAt),
+        last_uploaded_at: _base.getCurrentTime(),
+        last_downloaded_at: 0,
+        is_folder: false,
+        is_file: true,
+        download_in_progress: 0,
+        upload_in_progress: 0,
+      });
 
-        // Add an event log
-        await eventLogModel.create({
-          account_id: account.id,
-          type: eventType.UPLOAD_FILE,
-          description: `Uploaded File: ${filePath} to ${account.instance_url}`
-        });
-      })
-      .catch(async (error) => {
-        // Ignore "duplicate" status codes
-        if (error.statusCode == 409) {
-          // In case of duplicate error, we will update the file modified date to the db so that it does not try to update next time
-          await nodeModel.update({
-            file_update_at: _base.getFileModifiedTime(filePath)
-          }, {
-              where: {
-                account_id: account.id,
-                file_path: filePath
-              }
-            });
-        }
-        // If the file could be uploaded for some reason, we will delete the record so that the uploader can reintiate the transfer later
-        await nodeModel.destroy({
-          where: {
-            account_id: account.id,
-            file_path: filePath
-          }
-        });
-      })
+      console.log(`Done uploading file: ${filePath} to ${account.instance_url}`);
+
+      // Add an event log
+      await eventLogModel.create({
+        account_id: account.id,
+        type: eventType.UPLOAD_FILE,
+        description: `Uploaded File: ${filePath} to ${account.instance_url}`
+      });
+
+      return true;
+
+    } catch (error) {
+      // Ignore "duplicate" status codes
+      if (error.statusCode == 409) {
+        // In case of duplicate error, we will update the file modified date to the db so that it does not try to update next time
+        await nodeModel.update({
+          file_update_at: _base.getFileModifiedTime(filePath)
+        }, {
+            where: {
+              account_id: account.id,
+              file_path: filePath
+            }
+          });
+      }
+
+      return false;
+    }
   }
 
   return;
